backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "FastAPI backend for the Todo application"
dependencies = [
    "fastapi",
    "sqlmodel",
    "uvicorn[standard]",
    "psycopg2-binary",
    "python-dotenv",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
]
requires-python = ">=3.13"

[build-system]
requires = ["uv>=0.1.0", "uv-build"]
build-backend = "uv_build"

[tool.uv]
dev-dependencies = [] # No dev dependencies for this backend's pyproject.toml yet

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I"] # Error, Flake8, Warnings, Isort

[tool.black]
line-length = 88
target-version = ["py313"]
